+++ src/profiler_caliper.f90
@@ -37,7 +37,9 @@ module decomp_2d_profiler
              decomp_profiler_prep, &
              decomp_profiler_log, &
              decomp_profiler_start, &
-             decomp_profiler_end
+             decomp_profiler_end, &
+             decomp_profiler_pause, &
+             decomp_profiler_resume
 
    ! Generic interface to initialize the profiler
    interface decomp_profiler_init
@@ -68,6 +70,16 @@ module decomp_2d_profiler
    interface decomp_profiler_end
       module procedure decomp_profiler_end_char
    end interface decomp_profiler_end
+
+   ! Generic interface to pause profiling
+   interface decomp_profiler_pause
+      module procedure decomp_profiler_pause_noarg
+   end interface decomp_profiler_pause
+
+   ! Generic interface to resume profiling
+   interface decomp_profiler_resume
+      module procedure decomp_profiler_resume_noarg
+   end interface decomp_profiler_resume
 
 contains
 
@@ -177,6 +189,28 @@ contains
    end subroutine decomp_profiler_end_char
 
    !
+   ! Pause profiling operations
+   !
+   subroutine decomp_profiler_pause_noarg()
+
+      implicit none
+
+      call profiler_pause_or_resume(.true.)
+
+   end subroutine decomp_profiler_pause_noarg
+
+   !
+   ! Resume profiling operations
+   !
+   subroutine decomp_profiler_resume_noarg()
+
+      implicit none
+
+      call profiler_pause_or_resume(.false.)
+
+   end subroutine decomp_profiler_resume_noarg
+
+   !
    ! Not frequent, but the manager can produce an error
    !
    subroutine manager_error(manager)
@@ -198,5 +232,36 @@ contains
       end if
 
    end subroutine manager_error
+
+   !
+   ! Pause or resume profiling operations
+   !
+   subroutine profiler_pause_or_resume(prof_pause)
+
+      implicit none
+
+      ! Arugment
+      logical, intent(in) :: prof_pause
+
+      ! Local variable
+      logical, save, dimension(4) :: store
+
+      if (prof_pause) then
+         store(1) = decomp_profiler_transpose
+         store(2) = decomp_profiler_io
+         store(3) = decomp_profiler_fft
+         store(4) = decomp_profiler_d2d
+         decomp_profiler_transpose = .false.
+         decomp_profiler_io = .false.
+         decomp_profiler_fft = .false.
+         decomp_profiler_d2d = .false.
+      else
+         decomp_profiler_transpose = store(1)
+         decomp_profiler_io = store(2)
+         decomp_profiler_fft = store(3)
+         decomp_profiler_d2d = store(4)
+      end if
+
+   end subroutine profiler_pause_or_resume
 
 end module decomp_2d_profiler